tests/wpt/web-platform-tests/css/css-transitions/parsing/transition-property-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Transitions: parsing transition-property with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-transitions/#propdef-transition-property">
<meta name="assert" content="transition-property supports the full grammar 'none | <single-transition-property> #'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("transition-property", 'none');
test_valid_value("transition-property", 'all');
test_valid_value("transition-property", 'one');
test_valid_value("transition-property", 'one-two-three');
test_valid_value("transition-property", 'one, two, three');
test_valid_value("transition-property", 'width, all');
</script>
</body>
</html>
